Move Event into probe pkg
The events package only contains the Event type. This is only used in
the context of a probe, so move it to the probe package.
