+++ internal/pkg/instrumentation/bpf/google.golang.org/grpc/server/probe.go
@@ -21,6 +21,7 @@ import (
 	"os"
 
 	"go.opentelemetry.io/auto/internal/pkg/instrumentation/bpffs"
+	"go.opentelemetry.io/auto/internal/pkg/instrumentation/probe"
 
 	"github.com/cilium/ebpf"
 	"github.com/cilium/ebpf/link"
@@ -34,7 +35,6 @@ import (
 
 	"go.opentelemetry.io/auto/internal/pkg/inject"
 	"go.opentelemetry.io/auto/internal/pkg/instrumentation/context"
-	"go.opentelemetry.io/auto/internal/pkg/instrumentation/events"
 	"go.opentelemetry.io/auto/internal/pkg/instrumentation/utils"
 	"go.opentelemetry.io/auto/internal/pkg/process"
 	"go.opentelemetry.io/auto/internal/pkg/structfield"
@@ -184,7 +184,7 @@ func (g *Probe) Load(exec *link.Executable, target *process.TargetDetails) error
 }
 
 // Run runs the events processing loop.
-func (g *Probe) Run(eventsChan chan<- *events.Event) {
+func (g *Probe) Run(eventsChan chan<- *probe.Event) {
 	var event Event
 	for {
 		record, err := g.eventsReader.Read()
@@ -210,7 +210,7 @@ func (g *Probe) Run(eventsChan chan<- *events.Event) {
 	}
 }
 
-func (g *Probe) convertEvent(e *Event) *events.Event {
+func (g *Probe) convertEvent(e *Event) *probe.Event {
 	method := unix.ByteSliceToString(e.Method[:])
 
 	sc := trace.NewSpanContext(trace.SpanContextConfig{
@@ -232,7 +232,7 @@ func (g *Probe) convertEvent(e *Event) *events.Event {
 		pscPtr = nil
 	}
 
-	return &events.Event{
+	return &probe.Event{
 		Library:   g.LibraryName(),
 		Name:      method,
 		Kind:      trace.SpanKindServer,
