+++ internal/pkg/instrumentation/bpf/net/http/client/probe.go
@@ -32,7 +32,7 @@ import (
 	"go.opentelemetry.io/auto/internal/pkg/inject"
 	"go.opentelemetry.io/auto/internal/pkg/instrumentation/bpffs"
 	"go.opentelemetry.io/auto/internal/pkg/instrumentation/context"
-	"go.opentelemetry.io/auto/internal/pkg/instrumentation/events"
+	"go.opentelemetry.io/auto/internal/pkg/instrumentation/probe"
 	"go.opentelemetry.io/auto/internal/pkg/process"
 	"go.opentelemetry.io/auto/internal/pkg/structfield"
 )
@@ -148,7 +148,7 @@ func (h *Probe) Load(exec *link.Executable, target *process.TargetDetails) error
 }
 
 // Run runs the events processing loop.
-func (h *Probe) Run(eventsChan chan<- *events.Event) {
+func (h *Probe) Run(eventsChan chan<- *probe.Event) {
 	var event Event
 	for {
 		record, err := h.eventsReader.Read()
@@ -174,7 +174,7 @@ func (h *Probe) Run(eventsChan chan<- *events.Event) {
 	}
 }
 
-func (h *Probe) convertEvent(e *Event) *events.Event {
+func (h *Probe) convertEvent(e *Event) *probe.Event {
 	method := unix.ByteSliceToString(e.Method[:])
 	path := unix.ByteSliceToString(e.Path[:])
 
@@ -197,7 +197,7 @@ func (h *Probe) convertEvent(e *Event) *events.Event {
 		pscPtr = nil
 	}
 
-	return &events.Event{
+	return &probe.Event{
 		Library:     h.LibraryName(),
 		Name:        path,
 		Kind:        trace.SpanKindClient,
