+++ internal/pkg/instrumentation/probe/event.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package events
+package probe
 
 import (
 	"go.opentelemetry.io/otel/attribute"
